Separate IEditorConfigOptionMappingService from IGlobalOptionService (#61899)
* Separate IEditorConfigOptionMappingService from IGlobalOptionService

* Move to separate file

* Cleanup GlobalOptionService

* Remove dead code
