+++ src/EditorFeatures/Core/Logging/FunctionIdOptions.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Collections.Concurrent;
 using Microsoft.CodeAnalysis.Options;
 using Roslyn.Utilities;
+using System.Collections.Generic;
 
 namespace Microsoft.CodeAnalysis.Internal.Log
 {
@@ -25,13 +26,18 @@ namespace Microsoft.CodeAnalysis.Internal.Log
                 storageLocation: new LocalUserProfileStorageLocation(@"Roslyn\Internal\Performance\FunctionId\" + name));
         }
 
+        private static IEnumerable<FunctionId> GetFunctionIds()
+            => Enum.GetValues(typeof(FunctionId)).Cast<FunctionId>();
+
+        public static IEnumerable<IOption> GetOptions()
+            => GetFunctionIds().Select(GetOption);
+
         public static Option2<bool> GetOption(FunctionId id)
             => s_options.GetOrAdd(id, s_optionCreator);
 
         public static Func<FunctionId, bool> CreateFunctionIsEnabledPredicate(IGlobalOptionService globalOptions)
         {
-            var functionIds = Enum.GetValues(typeof(FunctionId)).Cast<FunctionId>();
-            var functionIdOptions = functionIds.ToDictionary(id => id, id => globalOptions.GetOption(GetOption(id)));
+            var functionIdOptions = GetFunctionIds().ToDictionary(id => id, id => globalOptions.GetOption(GetOption(id)));
             return functionId => functionIdOptions[functionId];
         }
     }
