+++ src/VisualStudio/Core/Impl/Options/AbstractOptionPreviewViewModel.cs
@@ -139,7 +139,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.Options
 
             var document = project.AddDocument("document", SourceText.From(text, Encoding.UTF8));
             var fallbackFormattingOptions = _globalOptions.GetSyntaxFormattingOptions(document.Project.LanguageServices);
-            var optionService = workspace.Services.GetRequiredService<ILegacyWorkspaceOptionService>();
+            var optionService = workspace.Services.GetRequiredService<IEditorConfigOptionMappingService>();
             var configOptions = OptionStore.GetOptions().AsAnalyzerConfigOptions(optionService, document.Project.Language);
             var formattingService = document.GetRequiredLanguageService<ISyntaxFormattingService>();
             var formattingOptions = formattingService.GetFormattingOptions(configOptions, fallbackFormattingOptions);
