+++ src/Workspaces/CoreTestUtilities/OptionsCollection.cs
@@ -76,7 +76,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.CodeActions
 
         public AnalyzerConfigOptions ToAnalyzerConfigOptions(HostLanguageServices languageServices)
         {
-            var optionService = languageServices.WorkspaceServices.GetRequiredService<ILegacyWorkspaceOptionService>();
+            var optionService = languageServices.WorkspaceServices.GetRequiredService<IEditorConfigOptionMappingService>();
             return ToOptionSet().AsAnalyzerConfigOptions(optionService, languageServices.Language);
         }
